chore(rolldown_plugin_vite_html): tweak `utils.rs` (#6460)
Separated the code in `utils.rs` into multiple files for better organization.
