+++ crates/rolldown_plugin_vite_html/src/utils/constant.rs
@@ -1,0 +1,6 @@
+use std::sync::LazyLock;
+
+use regex::Regex;
+
+pub static INLINE_IMPORT: LazyLock<Regex> =
+  LazyLock::new(|| Regex::new(r#"\bimport\s*\(("(?:[^"\\]|\\.)*"|'(?:[^'\\]|\\.)*')\)"#).unwrap());
