+++ crates/rolldown_plugin_vite_html/src/utils/mod.rs
@@ -1,0 +1,8 @@
+mod helpers;
+mod plugin_ext;
+mod script_inline_import_visistor;
+
+pub mod constant;
+
+pub use helpers::{is_excluded_url, overwrite_check_public_file};
+pub use script_inline_import_visistor::ScriptInlineImportVisitor;
