+++ crates/rolldown_plugin_vite_html/src/utils/plugin_ext.rs
@@ -1,0 +1,33 @@
+use std::borrow::Cow;
+
+use rolldown_plugin_utils::constants::{HTMLProxyMap, HTMLProxyMapItem};
+
+use crate::ViteHtmlPlugin;
+
+impl ViteHtmlPlugin {
+  pub fn get_base_in_html(&self, url_relative_path: &str) -> Cow<'_, str> {
+    if self.url_base.is_empty() || self.url_base == "./" {
+      let count = url_relative_path.matches('/').count();
+      Cow::Owned(if count == 0 { "./".to_owned() } else { "../".repeat(count) })
+    } else {
+      Cow::Borrowed(self.url_base.as_ref())
+    }
+  }
+
+  #[inline]
+  pub fn add_to_html_proxy_cache(
+    &self,
+    ctx: &rolldown_plugin::PluginContext,
+    file: String,
+    index: usize,
+    result: HTMLProxyMapItem,
+  ) {
+    ctx
+      .meta()
+      .get_or_insert_default::<HTMLProxyMap>()
+      .inner
+      .entry(file)
+      .or_default()
+      .insert(index, result);
+  }
+}
